fix methods

+++ src/main/java/com/tp_construccion/crud/service/IAutoServiceImpl.java
@@ -2,18 +2,20 @@ package com.tp_construccion.crud.service;
 
 import com.tp_construccion.crud.entities.Auto;
 import com.tp_construccion.crud.repository.AutoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class IAutoServiceImpl implements IAutoService {
 
-    @Autowired
-    private AutoRepository autoRepository;
+    private final AutoRepository autoRepository;
+
+    public IAutoServiceImpl(AutoRepository autoRepository) {
+        this.autoRepository = autoRepository;
+    }
 
     @Override
     public Auto crear(Auto auto) {
@@ -27,26 +29,26 @@ public class IAutoServiceImpl implements IAutoService {
 
     @Override
     public Auto actualizar(Long id, double nuevoPrecio) {
-        Optional<Auto> auto = this.autoRepository.findById(id);
-        if (auto.isPresent()) {
-           Auto tmp = auto.get();
-           tmp.setPrecio(nuevoPrecio);
-           return crear(tmp);
-        }
-        return null;
+        return this.autoRepository.findById(id)
+                .map(auto -> {
+                    auto.setPrecio(nuevoPrecio);
+                    return autoRepository.save(auto);
+                })
+                .orElseThrow(() -> new EntityNotFoundException("El auto con ID " + id + " no existe."));
     }
 
     @Override
     public boolean eliminar(Long id) {
-        this.autoRepository.deleteById(id);
-        return true;
+        if (this.autoRepository.existsById(id)) {
+            this.autoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
     public List<Auto> obtenerTodos() {
-        List<Auto> autos = new ArrayList<>();
-        this.autoRepository.findAll().forEach(autos::add);
-        return autos;
+        return this.autoRepository.findAll();
     }
 
     @Override
